requirements.txt
flask
torch
transformers
pillow
numpy
faiss-cpu
tqdm
gunicorn
psutil
